+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Create hacky script to run the REPL from inside Blender
   - Add `vb/obj-pointer` to be used for Blender-embedded Vybe
   - Add functions `vybe.blender` to set GLTF model export path and baking
+  - Sync working, can ignore entities by their blender identifiers (using `vb/entities-ignore!`)
 
 ## v0.7.522